CRISP Plugin: fix for Update Rate [ms] ui control with new MS2000 firmware versioning convention 9.50+

+++ plugins/CRISP/src/main/java/com/asiimaging/crisp/data/Defaults.java
@@ -2,7 +2,7 @@
  * Project: ASI CRISP Control
  * License: BSD 3-clause, see LICENSE.md
  * Author: Brandon Simpson (brandon@asiimaging.com)
- * Copyright (c) 2014-2021, Applied Scientific Instrumentation
+ * Copyright (c) 2014-2024, Applied Scientific Instrumentation
  */
 
 package com.asiimaging.crisp.data;
@@ -17,8 +17,8 @@ public final class Defaults {
    public static final int NUM_AVERAGES = 0;
    public static final int LED_INTENSITY = 50;
    public static final int UPDATE_RATE_MS = 10;
-   public static final float LOCK_RANGE = 1.0f;
-   public static final float OBJECTIVE_NA = 0.65f;
+   public static final double LOCK_RANGE = 1.0;
+   public static final double OBJECTIVE_NA = 0.65;
 
    // plugin settings
    public static final int POLL_RATE_MS = 250;
